pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crypto-trading-bot"
version = "0.1.0"
description = "AI-powered crypto trading bot with sentiment analysis and learning feedback loop"
requires-python = ">=3.11"
dependencies = [
    "ccxt>=4.0",
    "pandas>=2.0",
    "numpy>=1.24",
    "ta>=0.11",
    "vaderSentiment>=3.3",
    "aiohttp>=3.9",
    "sqlalchemy>=2.0",
    "rich>=13.0",
    "python-dotenv>=1.0",
    "click>=8.1",
    "scikit-learn>=1.3",
    "asyncpraw>=7.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
]

[project.scripts]
tradingbot = "src.cli:cli"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
